rename example app package name
