+++ in_app_review/example/android/app/src/main/kotlin/dev/britannio/in_app_review_demo/MainActivity.kt
@@ -1,4 +1,4 @@
-package dev.britannio.in_app_review_example
+package dev.britannio.in_app_review_demo
 
 import io.flutter.embedding.android.FlutterActivity
